Remove @style from video
It was there in the deprecated repository, but malformed; when I
imported the data, I fixed the malformed attribute in an attempt to
fix the styling of the video -- turns out, the styling of the video
actually came from the CSS I failed to import.

So the @style never worked and never should have worked, the video
styling was managed by CSS only.